カスタムフックへ、関数を切り離スタメ
